add ocaml-variants.5.1.1+flambda2
This introduces the flambda2 in-development variant, along with
the various base packages required to compile it. These are secondary
packages for dune/menhir (required to bootstrap the development version
of the forked compiler), and also a new base-flambda2 package to

The only variation to package selection should be that the new
ocaml-secondary-compiler could be selected by existing ocamlfind-secondary
packages, but it should be fine as it is just a newer version of
the bootstrap OCaml used to build the ocamlfind binaries.