Add governance and policy documents to repo
Up to this point, the explicit documents of our policies and governance
contacts have been kept in the [GitHub
wiki](https://github.com/ocaml/opam-repository/wiki) associated with
this repository.

However, in [today's opam repo maintenance
meeting](https://github.com/ocaml/opam-repository/wiki/Meeting-notes#2024-11-28),
we resolved that it would be beneficial to move the documents detailing
the policy and governance of this repository into the repository itself.
This provides the following benefits over the current practice:

- The policy will be visible when the repo is checked out, making easy
  to consult locally, and at the site of the work it governs.
- We can suggest and manage changes to the policy via PRs, which will
  - make it easier to discuss changes while leaving a record of those
    discussions, and
  - ensure that proposed changes are visible to users of the repository,
    and available for their commentary and questions.

This change adds the content of the following Wiki pages:

- https://github.com/ocaml/opam-repository/wiki/Governance
- https://github.com/ocaml/opam-repository/wiki/Policies
- https://github.com/ocaml/opam-repository/wiki/Package-Archiving:-Plan