Improve the `x-reason-for-archival` values
Following discussion with @hannesm, @mseri in the opam repo maintainer
meeting today, we agreed that the strings used to record the maintenance
intent in the archiving policy were liable to cause confusion, since
some of them were phrased as a positive, when their use in
`x-reason-for-archival` was meant to indicate a negative.

This change attempts to eliminate that source of confusion.