Avoid referring to ez-conf-lib:exe
The `ez-conf-lib:exe` variable is set in `ez-conf-lib.config` generated
while building the `ez-conf-lib` package, and is set to the absolute
path to the `ez-conf-lib` script. This causes a problem if the opam
switch where `ez-conf-lib` is installed to is ever moved, as the
absolute path will no longer refer to the correct place. This also
creates problems for dune package management which moves all installed
artifacts of a package after running its install command.

The fix is to refer to the `ez-conf-lib` script directly by specifying
its path relative to the `lib` directory where it is installed.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>