conf-libcurl: Add mingw-w64-shims as a dependency
Installing `conf-libcurl` can trigger a rebuild of `mingw-w64-shims``,
which adjusts the `PATH` to include the `curl-config` script required
during the build. However, since `mingw-w64-shims` is not listed as a
direct dependency, it may be installed too late, leading to build
failures when `curl-config` is not found in the PATH.

This commit fixes the issue by declaring `mingw-w64-shims` as a direct
build dependency of `conf-libcurl`.